+++ cmd/version_test.go
@@ -1,0 +1,7 @@
+package cmd
+
+func ExampleVersionCmd_Run() {
+	cmd := &VersionCmd{}
+	_ = cmd.Run(nil, nil)
+	// Output: gphotos-uploader-cli v0.0.0
+}