+++ config/config.go
@@ -6,9 +6,9 @@ import (
 	"log"
 	"os"
 	"path"
+	"path/filepath"
 
 	"github.com/client9/xson/hjson"
-	"github.com/nmrshll/go-cp"
 
 	"github.com/gphotosuploader/gphotos-uploader-cli/utils/filesystem"
 )
@@ -39,7 +39,11 @@ func defaultSettings() *Config {
 // NewConfig returns a *Config with the default settings of the application.
 func NewConfig(dir string) *Config {
 	cfg := defaultSettings()
-	cfg.ConfigPath = filesystem.AbsolutePath(dir)
+	absPath, err := filesystem.AbsolutePath(dir)
+	if err != nil {
+		absPath = dir
+	}
+	cfg.ConfigPath = absPath
 
 	return cfg
 }
@@ -51,7 +55,7 @@ func (c *Config) Validate() error {
 
 	for i := range c.Jobs {
 		item := &c.Jobs[i] // we do that way to modify original object while iterating.
-		path, err := cp.AbsolutePath(item.SourceFolder)
+		path, err := filesystem.AbsolutePath(item.SourceFolder)
 		if err != nil {
 			return fmt.Errorf("invalid source folder. SourceFolder=%s, err=%s", item.SourceFolder, err)
 		}
@@ -138,35 +142,24 @@ func LoadConfig(dir string) (*Config, error) {
 	return cfg, nil
 }
 
-// InitConfig creates an example config file if it doesn't already exist.
-// If 'force' is set then we are going to remove config dir before creating it.
-func InitConfig(dir string, force bool) error {
+// InitConfigFile creates a config file with default settings.
+func InitConfigFile(dir string) error {
 	cfg := NewConfig(dir)
 
-	// if force, we should remove everything to start from the scratch.
-	if force {
-		err := os.RemoveAll(dir)
-		if err != nil {
-			return err
-		}
-	}
-
-	if _, err := os.Stat(cfg.ConfigPath); !os.IsNotExist(err) {
-		// directory already exist and forced was not set
-		return fmt.Errorf("config directory already exists, use '--force' to overwrite: path=%s", cfg.ConfigPath)
-	} else {
-		err := os.MkdirAll(cfg.ConfigPath, 0755)
-		if err != nil {
-			return fmt.Errorf("failed to create config directory: path=%s, err=%v", cfg.ConfigPath, err)
-		}
-	}
-
-	fh, err := os.Open(cfg.ConfigFile())
+	// Delete config & overwrite config
+	err := os.RemoveAll(dir)
 	if err != nil {
-		fh, err = os.Create(cfg.ConfigFile())
-		if err != nil {
-			return fmt.Errorf("failed to create config: file=%s, err=%v", cfg.ConfigFile(), err)
-		}
+		return err
+	}
+
+	err = os.MkdirAll(cfg.ConfigPath, 0755)
+	if err != nil {
+		return fmt.Errorf("failed to create config directory: path=%s, err=%v", cfg.ConfigPath, err)
+	}
+
+	fh, err := os.Create(cfg.ConfigFile())
+	if err != nil {
+		return fmt.Errorf("failed to create config: file=%s, err=%v", cfg.ConfigFile(), err)
 	}
 	defer func() {
 		if err := fh.Close(); err != nil {
@@ -180,4 +173,19 @@ func InitConfig(dir string, force bool) error {
 	}
 
 	return fh.Sync()
+}
+
+// ConfigExists checks if a gphotos-uplaoder-cli configuration exists at a certain path
+func ConfigExists(path string) bool {
+	// Check config.hjson
+	cfgFile, err := filesystem.AbsolutePath(filepath.Join(path, DefaultConfigsPath))
+	if err != nil {
+		return false
+	}
+
+	if _, err := os.Stat(cfgFile); err == nil {
+		return true
+	}
+
+	return false // Config file found
 }