+++ config/config_test.go
@@ -15,25 +15,17 @@ func TestInitConfig(t *testing.T) {
 	dir := filepath.Join(os.TempDir(), fmt.Sprintf("gphotos-config.%d", time.Now().UnixNano()))
 
 	t.Run("TestInitConfigFile", func(t *testing.T) {
-		err := config.InitConfig(dir, true)
+		err := config.InitConfigFile(dir)
 		if err != nil {
 			t.Errorf("could not create init config file: %v", err)
 		}
 	})
-
 	defer func() {
 		err := os.RemoveAll(dir)
 		if err != nil {
 			t.Errorf("could not remove test config file (dir: %s): %v", dir, err)
 		}
 	}()
-
-	t.Run("TestInitConfigFileWithExistentFile", func(t *testing.T) {
-		err := config.InitConfig(dir, false)
-		if err == nil {
-			t.Error("an error creating an existent file was expected")
-		}
-	})
 }
 
 func TestInitAndLoadConfig(t *testing.T) {
@@ -41,12 +33,11 @@ func TestInitAndLoadConfig(t *testing.T) {
 	dir := filepath.Join(os.TempDir(), fmt.Sprintf("gphotos-config.%d", time.Now().UnixNano()))
 
 	t.Run("TestInitConfigFile", func(t *testing.T) {
-		err := config.InitConfig(dir, true)
+		err := config.InitConfigFile(dir)
 		if err != nil {
 			t.Errorf("could not create init config file: %v", err)
 		}
 	})
-
 	defer func() {
 		err := os.RemoveAll(dir)
 		if err != nil {
@@ -188,4 +179,47 @@ func createTestConfiguration() *config.Config {
 	}
 	c.Jobs = append(c.Jobs, job)
 	return c
+}
+
+func TestConfigExists(t *testing.T) {
+	dir := filepath.Join(os.TempDir(), fmt.Sprintf("gphotos-test.%d", time.Now().UnixNano()))
+	if err := os.RemoveAll(dir); err != nil {
+		t.Fatalf("no error was expected at this point: err=%s", err)
+	}
+
+	t.Run("TestNonExistingConfiguration", func(t *testing.T) {
+		if config.ConfigExists(dir) {
+			t.Errorf("config file exists. That was not expected: dir=%s", dir)
+		}
+	})
+
+	if err := os.MkdirAll(dir, os.ModePerm); err != nil {
+		t.Fatalf("no error was expected at this point: err=%s", err)
+	}
+	defer func() {
+		if err := os.RemoveAll(dir); err != nil {
+			t.Fatalf("could not remove test config folder: path=%s err=%s", dir, err)
+		}
+	}()
+
+	cfgFile := filepath.Join(dir, "config.hjson")
+	fh, err := os.Create(cfgFile)
+	if err != nil {
+		t.Fatalf("failed to create config: file=%s, err=%v", cfgFile, err)
+	}
+	defer func() {
+		if err := fh.Close(); err != nil {
+			t.Fatal(err)
+		}
+	}()
+
+	if _, err = fh.WriteString("testTest"); err != nil {
+		t.Fatalf("failed to write configuration: file=%s, err=%v", cfgFile, err)
+	}
+
+	t.Run("TestExistingConfiguration", func(t *testing.T) {
+		if !config.ConfigExists(dir) {
+			t.Errorf("config file doesn't exist. That was not expected: dir=%s", dir)
+		}
+	})
 }