+++ log/doc.go
@@ -1,0 +1,4 @@
+/*
+Package log implements a simple logger using Logrus. It implements several levels of logging and several targets also.
+*/
+package log