+++ utils/filesystem/filesystem_test.go
@@ -27,7 +27,7 @@ func TestAbsolutePath(t *testing.T) {
 		}
 
 		for _, test := range absolutePathInputs {
-			got := filesystem.AbsolutePath(test.in)
+			got, _ := filesystem.AbsolutePath(test.in)
 			if got != test.out {
 				t.Errorf("failed for '%s': expected '%v', got '%v'", test.in, test.out, got)
 			}
@@ -55,7 +55,7 @@ func TestAbsolutePath(t *testing.T) {
 		}
 
 		for _, test := range relativePathInputs {
-			got := filesystem.AbsolutePath(test.in)
+			got, _ := filesystem.AbsolutePath(test.in)
 			expected := path.Join(cwd, test.out)
 
 			if got != expected {
@@ -87,7 +87,7 @@ func TestAbsolutePath(t *testing.T) {
 		dir := usr.HomeDir
 
 		for _, test := range tildePathInputs {
-			got := filesystem.AbsolutePath(test.in)
+			got, _ := filesystem.AbsolutePath(test.in)
 			expected := path.Join(dir, test.out)
 
 			if got != expected {
@@ -104,6 +104,7 @@ func TestIsFile(t *testing.T) {
 	}{
 		{"testdata/file.txt", true},
 		{"testdata/folder", false},
+		{"testdata/non-existent-file", false},
 	}
 
 	for _, test := range objectsTest {
@@ -121,6 +122,7 @@ func TestIsDir(t *testing.T) {
 	}{
 		{"testdata/file.txt", false},
 		{"testdata/folder", true},
+		{"testdata/non-existent-dir", false},
 	}
 
 	for _, test := range objectsTest {